Add a configurable reply-to email
We use a Gmail account as our primary email, but we can't send emails
through Mandrill as a Gmail address – the sender email must be under domain
that's under our control as it has to be configured properly for Madrill.

We still want to make sure people can respond to our emails, so this commit
introduces a configurable contact email which is then used in the reply-to
field in outgoing emails.

A good alternative would be to not use reply-to, but rather configure the
Mandrill email so that it forwards any emails to the main email, but
that's somewhat more work.